# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Warehouse Management API
  description: |
    API for warehouse operations, stock, approvals, reporting, and exports.

    Request and response JSON property names use snake_case. Successful responses use the common
    envelope `{success, code, message, data, errors}`. Errors use the same envelope with `success=false`,
    `data=null`, and field errors keyed by snake_case property names. Bean validation failures return 422.

    Paginated list endpoints accept optional `page` (default 1), `size` (default 10, maximum 100),
    `sort` (endpoint-supported field), `order` (`asc` or `desc`, default `desc`), and `search`.
    Authentication uses a bearer JWT. Inbound, outbound, transfer, and adjustment records follow
    draft -> submit -> approve/reject workflows; approval and rejection require their corresponding
    permission, and creators cannot review their own records. Workflow conflicts return 409 with a
    stable error code such as `*_NOT_REVIEWABLE`, `*_NOT_EDITABLE`, or `INSUFFICIENT_STOCK`; a creator
    attempting self-review receives 403 with a `*_SELF_REVIEW_FORBIDDEN` code.
  version: 1.0.0
servers:
- url: http://localhost:8081
  description: Generated server url
...
paths:
...
  /api/v1/stock-transfers/{id}:
    get:
      tags:
      - Stock transfers
      operationId: get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStockTransferResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
    put:
      tags:
      - Stock transfers
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockTransferUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStockTransferResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Stock transfers
      operationId: delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
...
  /api/v1/stock-transfers:
    get:
      tags:
      - Stock transfers
      operationId: list_2
      parameters:
      - name: page
        in: query
        description: One-based page number.
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          description: One-based page number.
          minimum: 1
      - name: size
        in: query
        description: Number of records per page (1-100).
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          description: Number of records per page (1-100).
          maximum: 100
          minimum: 1
      - name: sort
        in: query
        description: Field used to sort the results.
        required: false
        schema:
          type: string
          default: created_at
          description: Field used to sort the results.
      - name: order
        in: query
        description: "Sort direction: asc or desc."
        required: false
        schema:
          type: string
          default: desc
          description: "Sort direction: asc or desc."
          enum:
          - asc
          - desc
          pattern: (?i)^(asc|desc)$
      - name: search
        in: query
        description: Optional text filter.
        required: false
        schema:
          type: string
          description: Optional text filter.
          maxLength: 100
          minLength: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseStockTransferListItemResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
    post:
      tags:
      - Stock transfers
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockTransferCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStockTransferResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
  /api/v1/stock-transfers/{id}/submit:
    post:
      tags:
      - Stock transfers
      operationId: submit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStockTransferResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
  /api/v1/stock-transfers/{id}/reject:
    post:
      tags:
      - Stock transfers
      operationId: reject
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockTransferRejectRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStockTransferResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
  /api/v1/stock-transfers/{id}/approve:
    post:
      tags:
      - Stock transfers
      operationId: approve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStockTransferResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
...
  /api/v1/stock-adjustments/{id}/approve:
    post:
      tags:
      - Stock adjustments
      operationId: approve_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStockAdjustmentResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
...
  /api/v1/outbounds/{id}/approve:
    post:
      tags:
      - Outbounds
      operationId: approve_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseOutboundResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
...
  /api/v1/inbounds/{id}/approve:
    post:
      tags:
      - Inbounds
      operationId: approve_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseInboundResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    StockTransferUpdateRequest:
      type: object
      properties:
        notes:
          type: string
          maxLength: 5000
          minLength: 0
        items:
          type: array
          items:
            $ref: "#/components/schemas/StockTransferItemRequest"
          minItems: 1
      required:
      - items
    ApiResponseStockTransferResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/StockTransferResponse"
        errors:
          type: object
          additionalProperties:
            type: string
...
    StockTransferResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        warehouse_id:
          type: string
          format: uuid
        reference_number:
          type: string
        status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
        notes:
          type: string
        rejection_note:
          type: string
        submitted_at:
          type: string
          format: date-time
        reviewed_at:
          type: string
          format: date-time
        created_by_user_id:
          type: string
          format: uuid
        reviewed_by_user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/StockTransferItemResponse"
...
    ApiResponseStockAdjustmentResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/StockAdjustmentResponse"
        errors:
          type: object
          additionalProperties:
            type: string
...
    ApiResponseOutboundResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/OutboundResponse"
        errors:
          type: object
          additionalProperties:
            type: string
...
    ApiResponseInboundResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/InboundResponse"
        errors:
          type: object
          additionalProperties:
            type: string
...
    StockTransferCreateRequest:
      type: object
      properties:
        warehouse_id:
          type: string
          format: uuid
        notes:
          type: string
          maxLength: 5000
          minLength: 0
        items:
          type: array
          items:
            $ref: "#/components/schemas/StockTransferItemRequest"
          minItems: 1
      required:
      - items
      - warehouse_id
    StockTransferRejectRequest:
      type: object
      properties:
        rejection_note:
          type: string
          maxLength: 5000
          minLength: 0
      required:
      - rejection_note
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data: {}
        errors:
          type: object
          additionalProperties:
            type: string
...
    ApiResponsePageResponseStockTransferListItemResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/PageResponseStockTransferListItemResponse"
        errors:
          type: object
          additionalProperties:
            type: string
...
    ApiErrorEnvelope:
      type: object
      properties:
        success:
          default: false
        code:
          type: integer
          format: int32
        message:
          type: string
        data: {}
        errors:
          type: object
          additionalProperties:
            type: string
      required:
      - code
      - data
      - errors
      - message
      - success
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT